backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
torch
transformers
pdfplumber
faiss-cpu
sentence-transformers
